Using the module generator scripts for completeness
