+++ docs/modules/mockserver.md
@@ -1,10 +1,10 @@
 # MockServer
 
-Since testcontainers-go <a href="https://github.com/testcontainers/testcontainers-go/releases/tag/v0.24.0"><span class="tc-version">:material-tag: v0.24.0</span></a>
+Not available until the next release of testcontainers-go <a href="https://github.com/testcontainers/testcontainers-go"><span class="tc-version">:material-tag: main</span></a>
 
 ## Introduction
 
-Mock Server can be used to mock HTTP services by matching requests against user-defined expectations.
+The Testcontainers module for MockServer.
 
 ## Adding this module to your project dependencies
 
@@ -17,7 +17,7 @@ go get github.com/testcontainers/testcontainers-go/modules/mockserver
 ## Usage example
 
 <!--codeinclude-->
-[Creating a MockServer container](../../modules/mockserver/mockserver_test.go) inside_block:runMockServerContainer
+[Creating a MockServer container](../../modules/mockserver/examples_test.go) inside_block:runMockServerContainer
 <!--/codeinclude-->
 
 ## Module reference
@@ -40,26 +40,19 @@ When starting the MockServer container, you can pass options in a variadic way t
 If you need to set a different MockServer Docker image, you can use `testcontainers.WithImage` with a valid Docker image
 for MockServer. E.g. `testcontainers.WithImage("mockserver/mockserver")`.
 
+{% include "../features/common_functional_options.md" %}
+
 ### Container Methods
 
 The MockServer container exposes the following methods:
 
-#### GetHost
+#### URL
 
-The `GetHost` method returns the host to connect to the MockServer container.
-
-It can be use to configure a MockServer client (`github.com/BraspagDevelopers/mock-server-client`), e.g.:
-
-<!--codeinclude-->
-[Using Gethost with the MockServer client](../../modules/mockserver/mockserver_test.go) inside_block:connectToMockServer
-<!--/codeinclude-->
-
-#### GetPort
-
-The `GetPort` method returns the port to connect to the MockServer container.
+The `URL` method returns the url string to connect to the MockServer container.
+It returns a string with the format `http://<host>:<port>`.
 
 It can be use to configure a MockServer client (`github.com/BraspagDevelopers/mock-server-client`), e.g.:
 
 <!--codeinclude-->
-[Using GetPort with the MockServer client](../../modules/mockserver/mockserver_test.go) inside_block:connectToMockServer
+[Using URL with the MockServer client](../../modules/mockserver/examples_test.go) inside_block:connectToMockServer
 <!--/codeinclude-->
