+++ mkdocs.yml
@@ -74,6 +74,7 @@ nav:
         - modules/kafka.md
         - modules/localstack.md
         - modules/mariadb.md
+        - modules/mockserver.md
         - modules/mongodb.md
         - modules/mssql.md
         - modules/mysql.md
